requirements.txt
fastapi
uvicorn[standard]
prometheus-client
python-dotenv
